+++ src/extension.ts
@@ -113,6 +113,11 @@ export async function activate(context: vscode.ExtensionContext): Promise<void> 
 	const tutorial3Uri = vscode.Uri.parse(`file://${tutorial3Path}`);
 	await registerTutorial("HelloWorld with JavaScript in Three Steps", tutorial3Uri.fsPath, DEFAULT_TUTORIAL_CATEGORY);
 
+	// register the didact tutorial
+	const tutorial4Path = path.join(context.extensionPath, './demos/markdown/tutorial/tutorial.didact.md');
+	const tutorial4Uri = vscode.Uri.parse(`file://${tutorial4Path}`);
+	await registerTutorial("Writing Your First Didact Tutorial", tutorial4Uri.fsPath, DEFAULT_TUTORIAL_CATEGORY);
+
 	// create the view
 	createIntegrationsView();