+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to the "vscode-didact" extension will be documented in this 
  - Adding extension setting to enable opening the default Didact file at startup. [#144](https://github.com/redhat-developer/vscode-didact/issues/144)
  - Switching to xmldom for heading parsing and better Typescript 4 support
  - Adding ability to open a Didact file in a different column by default
+ - Adding simple history to enable stepping forward/back through last few Didact tutorials [#102](https://github.com/redhat-developer/vscode-didact/issues/102)
 
 ## 0.1.16